Set TextViews to data from received object

+++ app/src/main/java/com/marwaeltayeb/weatherforecast/model/Sys.kt
@@ -15,8 +15,8 @@ data class Sys(
     val country: String,
 
     @SerializedName("sunrise")
-    val sunrise: Int,
+    val sunrise: Long,
 
     @SerializedName("sunset")
-    val sunset: Int
+    val sunset: Long
 ): Serializable
